pyproject.toml
[tool.poetry]
name = "flo"
version = "0.1.0"
description = "A easy way to create AI agents"
authors = ["vizsatiz <vishnu@rootfor.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
langchain = "0.2.1"
langgraph = "0.0.55"
ipython = "8.24.0"
httpx = "0.27.0"
langchain-community = "0.2.0"
langchain-experimental = "0.0.59"
google-cloud-bigquery = "^3.23.1"
openai = "1.30.5"
langchain-openai = "^0.1.8"
pillow = "^10.3.0"
langchain-mongodb = "^0.1.5"
langchain-mistralai = "^0.1.7"
langchain-chroma = "^0.1.1"


[tool.poetry.group.dev.dependencies]
langchainhub = "^0.1.17"
torch = "2.0.0"
sentence-transformers = "^3.0.0"
pypdf = "^4.2.0"
lark = "^1.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
